Tour Model + Db Migration
Description: Added a Tour Model that will serve as the structure for loading each entry of tours available for booking. The Tour Model also uses a Foreign Key that will correctly reference to the appriopriate Identity User that has created the tour.
Relevant Db Migrations to register the Tour Model with appriopriate Foreign Key reference to AspNetUser in the Database has been performed.