Send Personal Message Function
Description: Basic Implementation of Sending a personal message to another user of the site using a Message Viewmodel that will be created from the Compose Form and sent to the controller for processing to insert an an actual Message model into the Db. The future plans for this is to fully convert this function into an AJAX based web form instead of a ViewModel and also to implement Typeahead feature to query all the possible users in the DB and allow the user to select a valid user before sending the message.

Added:
  - MessagesController
      - Compose()
      - Send() [HTTPPost]

ViewModel:
  - SendMessageFormViewModel

Views --> Messages:
  - Compose