[Message] Mark Message as Read
Description: When a user now clicks on a View Message Button in their inbox, a HTTP Put request gets sent to the messages API controller to add a timestamp to a message if the DateRead attribute is null, thus marking the message as read. This is necessary as it attributes to the count of unread messages in the user summary when they visit the homepage.

Added:
 - Controller:
  -MessagesController [API]
    - ReadMessage(id)

 - Views/Messages/Inbox
   - Ajax call function added to on click View Message Button.