[User] Upcoming Bookings View Rework (BETA)
Description: Changed the View of the existing upcoming bookings which was a DataTable that shows all of the users upcoming booked tours into a card format similar to the Available Tours Rework. Users can still see the relevant details related to their bookings and perform related actions to each booking.

Added:
 - Views/Tours:
    - UpcomingBookingsBeta