[Tours] Available Tours Prevent Overbooking
Description: Available Tours will now no longer show upcoming tours that have been fully booked. If a page has been pre-loaded with a tour that was concurrently fully booked while the user was browsing and should the user decide to book the same tour, the booking request will fail and the user will be alerted that the tour is full.